# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.AccountStatus:
    properties:
      account:
        $ref: '#/definitions/models.AccountResponse'
      is_configured:
        type: boolean
      recommendations:
        items:
          type: string
        type: array
      warnings:
        items:
          type: string
        type: array
    type: object
...
  controllers.TaxAccountStatusResponse:
    properties:
      cogs:
        $ref: '#/definitions/controllers.AccountStatus'
      configuration_id:
        type: integer
      health_score:
        description: 0-100
        type: integer
      inventory:
        $ref: '#/definitions/controllers.AccountStatus'
      is_fully_configured:
        type: boolean
      last_updated:
        type: string
      missing_accounts:
        description: Overall status
        items:
          type: string
        type: array
      purchase_bank:
        $ref: '#/definitions/controllers.AccountStatus'
      purchase_cash:
        $ref: '#/definitions/controllers.AccountStatus'
      purchase_expense:
        $ref: '#/definitions/controllers.AccountStatus'
      purchase_input_vat:
        $ref: '#/definitions/controllers.AccountStatus'
      purchase_payable:
        allOf:
        - $ref: '#/definitions/controllers.AccountStatus'
        description: Purchase accounts status
      sales_bank:
        $ref: '#/definitions/controllers.AccountStatus'
      sales_cash:
        $ref: '#/definitions/controllers.AccountStatus'
      sales_output_vat:
        $ref: '#/definitions/controllers.AccountStatus'
      sales_receivable:
        allOf:
        - $ref: '#/definitions/controllers.AccountStatus'
        description: Sales accounts status
      sales_revenue:
        $ref: '#/definitions/controllers.AccountStatus'
      system_warnings:
        items:
          type: string
        type: array
      tax_payable:
        $ref: '#/definitions/controllers.AccountStatus'
      updated_by:
        $ref: '#/definitions/models.UserResponse'
      withholding_tax21:
        allOf:
        - $ref: '#/definitions/controllers.AccountStatus'
        description: Optional accounts
      withholding_tax23:
        $ref: '#/definitions/controllers.AccountStatus'
      withholding_tax25:
        $ref: '#/definitions/controllers.AccountStatus'
    type: object
...
  models.AccountResponse:
    properties:
      balance:
        example: 5000000
        type: number
      category:
        example: current_asset
        type: string
      code:
        example: 1100-001
        type: string
      created_at:
        example: "2024-01-15T09:30:00Z"
        type: string
      description:
        example: Kas untuk keperluan operasional harian
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      is_header:
        example: false
        type: boolean
      level:
        example: 3
        type: integer
      name:
        example: Kas Kecil
        type: string
      parent_id:
        example: 1
        type: integer
      type:
        example: asset
        type: string
      updated_at:
        example: "2024-01-15T09:30:00Z"
        type: string
    type: object
...
  models.TaxAccountSettingsCreateRequest:
    properties:
      apply_to_all_companies:
        description: Configuration flags
        type: boolean
      cogs_account_id:
        type: integer
      inventory_account_id:
        description: Inventory Account (optional)
        type: integer
      notes:
        type: string
      purchase_bank_account_id:
        type: integer
      purchase_cash_account_id:
        type: integer
      purchase_expense_account_id:
        type: integer
      purchase_input_vat_account_id:
        type: integer
      purchase_payable_account_id:
        description: Purchase Account Configuration
        type: integer
      sales_bank_account_id:
        type: integer
      sales_cash_account_id:
        type: integer
      sales_output_vat_account_id:
        type: integer
      sales_receivable_account_id:
        description: Sales Account Configuration
        type: integer
      sales_revenue_account_id:
        type: integer
      tax_payable_account_id:
        type: integer
      withholding_tax21_account_id:
        description: Other Tax Accounts (optional)
        type: integer
      withholding_tax23_account_id:
        type: integer
      withholding_tax25_account_id:
        type: integer
    required:
    - purchase_bank_account_id
    - purchase_cash_account_id
    - purchase_expense_account_id
    - purchase_input_vat_account_id
    - purchase_payable_account_id
    - sales_bank_account_id
    - sales_cash_account_id
    - sales_output_vat_account_id
    - sales_receivable_account_id
    - sales_revenue_account_id
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        example: "2024-01-15T09:30:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      name:
        description: For compatibility with tax_account_settings
        example: John Doe
        type: string
      role:
        example: admin
        type: string
      updated_at:
        example: "2024-01-15T09:30:00Z"
        type: string
      username:
        example: john_doe
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API untuk aplikasi sistem akuntansi yang komprehensif dengan fitur
    lengkap manajemen keuangan, inventory, sales, purchases, dan reporting.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sistema Akuntansi API
  version: "1.0"
paths:
...
  /api/v1/settings/tax-accounts/status:
    get:
      consumes:
      - application/json
      description: Get detailed status information about tax account configuration
        including validation and recommendations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.TaxAccountStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get tax account configuration status
      tags:
      - Tax Account Status
  /api/v1/settings/tax-accounts/validate:
    get:
      consumes:
      - application/json
      description: Validate if a specific account is suitable for a particular tax
        account role
      parameters:
      - description: Account ID to validate
        in: query
        name: account_id
        required: true
        type: integer
      - description: Role to validate for (sales_receivable, sales_cash, etc.)
        in: query
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Validate account selection
      tags:
      - Tax Account Status
    post:
      consumes:
      - application/json
      description: Validate the tax account configuration before saving
      parameters:
      - description: Tax account settings data
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/models.TaxAccountSettingsCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Validate tax account configuration
      tags:
      - Tax Account Settings
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"